# protobuf: copy input data before decoding