@carl-mastrangelo 2016-08-12
Unexpectedly, there is a latency improvement on non-throughput bounded RPCs:

```
After:
Benchmark                         (direct)  (transport)    Mode     Cnt       Score     Error  Units
TransportBenchmark.unaryCall1024      true    INPROCESS  sample  843738    1636.493 ±  25.616  ns/op
TransportBenchmark.unaryCall1024      true        NETTY  sample  261953  152453.549 ± 410.259  ns/op
TransportBenchmark.unaryCall1024      true  NETTY_LOCAL  sample  358631  111337.198 ± 383.789  ns/op
TransportBenchmark.unaryCall1024      true       OKHTTP  sample  275771  144798.488 ± 444.639  ns/op
TransportBenchmark.unaryCall1024     false    INPROCESS  sample  997583   18374.480 ±  67.945  ns/op
TransportBenchmark.unaryCall1024     false        NETTY  sample  196664  203092.272 ± 599.238  ns/op
TransportBenchmark.unaryCall1024     false  NETTY_LOCAL  sample  243526  164007.563 ± 524.886  ns/op
TransportBenchmark.unaryCall1024     false       OKHTTP  sample  207957  192037.259 ± 666.798  ns/op

Before:
Benchmark                         (direct)  (transport)    Mode     Cnt       Score     Error  Units
TransportBenchmark.unaryCall1024      true    INPROCESS  sample  845992    1657.045 ±  35.495  ns/op
TransportBenchmark.unaryCall1024      true        NETTY  sample  248463  160749.033 ± 527.269  ns/op
TransportBenchmark.unaryCall1024      true  NETTY_LOCAL  sample  339227  117719.943 ± 428.632  ns/op
TransportBenchmark.unaryCall1024      true       OKHTTP  sample  260116  153515.303 ± 580.275  ns/op
TransportBenchmark.unaryCall1024     false    INPROCESS  sample  984239   18638.299 ±  78.800  ns/op
TransportBenchmark.unaryCall1024     false        NETTY  sample  180027  221856.332 ± 671.632  ns/op
TransportBenchmark.unaryCall1024     false  NETTY_LOCAL  sample  237268  168316.719 ± 679.528  ns/op
TransportBenchmark.unaryCall1024     false       OKHTTP  sample  195578  204207.931 ± 764.328  ns/op
```

More benchmarks to come.