@carl-mastrangelo 2016-08-13
Running with non bytestring based messages shows strong improvements too.  This was run with a proto with 1000 repeated int32s, linearly scaled from 0  to INTMAX:

```
Before:
Benchmark                       (direct)  (transport)    Mode      Cnt       Score      Error  Units
TransportBenchmark.nonBytesMsg      true    INPROCESS  sample  1121635    1213.588 ±   22.526  ns/op
TransportBenchmark.nonBytesMsg      true        NETTY  sample   146777  272127.649 ±  748.099  ns/op
TransportBenchmark.nonBytesMsg      true  NETTY_LOCAL  sample   195136  204689.484 ±  434.612  ns/op
TransportBenchmark.nonBytesMsg      true       OKHTTP  sample   123221  324022.729 ± 1135.778  ns/op
TransportBenchmark.nonBytesMsg     false    INPROCESS  sample   994066   19675.590 ±   63.081  ns/op
TransportBenchmark.nonBytesMsg     false        NETTY  sample   123667  322932.864 ±  974.805  ns/op
TransportBenchmark.nonBytesMsg     false  NETTY_LOCAL  sample   169397  235815.245 ±  511.067  ns/op
TransportBenchmark.nonBytesMsg     false       OKHTTP  sample   107508  371383.411 ± 1281.658  ns/op


After
Benchmark                       (direct)  (transport)    Mode      Cnt       Score     Error  Units
TransportBenchmark.nonBytesMsg      true    INPROCESS  sample  1138109    1182.735 ±  15.692  ns/op
TransportBenchmark.nonBytesMsg      true        NETTY  sample   232062  172099.083 ± 488.111  ns/op
TransportBenchmark.nonBytesMsg      true  NETTY_LOCAL  sample   350577  113907.426 ± 439.472  ns/op
TransportBenchmark.nonBytesMsg      true       OKHTTP  sample   195414  204393.604 ± 588.232  ns/op
TransportBenchmark.nonBytesMsg     false    INPROCESS  sample  1030020   19204.172 ±  47.569  ns/op
TransportBenchmark.nonBytesMsg     false        NETTY  sample   191721  208327.647 ± 528.368  ns/op
TransportBenchmark.nonBytesMsg     false  NETTY_LOCAL  sample   259517  153891.194 ± 514.530  ns/op
TransportBenchmark.nonBytesMsg     false       OKHTTP  sample   165339  241571.428 ± 733.777  ns/op
```

The approximate payload size is 4K, and exercises a different code path through proto.