@pgrosu 2016-08-13
Carl, this is great!  Here's another way of showing the impact of improvement of the two tests, in terms of percentage on the nanosecond score of the optimization:

![test1](https://cloud.githubusercontent.com/assets/6555937/17640771/8b82496c-60d5-11e6-9148-505cf54baadf.png)

![test2](https://cloud.githubusercontent.com/assets/6555937/17640772/8e8874c4-60d5-11e6-9346-103e172b16b7.png)

Hope it helps,
~p