@nmittler 2016-08-15
@carl-mastrangelo LGTM

@ejona86 want to take a look?