@carl-mastrangelo 2016-08-15
Results [before](https://github.com/grpc/grpc-java/files/418634/before.txt) and [after](https://github.com/grpc/grpc-java/files/418635/after.txt).  

Pretty much everything is better across the board.  Run with 10 forks.  There are still some puzzling things,  like things that shouldn't have improved.  I ran this on my unloaded system (each run takes about 8 hours) so I am not sure where the discrepancy is coming from.