@carl-mastrangelo 2016-08-17
backed out the cast, using byte buffer.