@carl-mastrangelo 2016-08-17
Tests fail because Compressed input streams lie about their available size (returning 1).  InputStream is a bad api.