@pgrosu 2016-08-17
@carl-mastrangelo Sorry man, yeah it's not a precise function.  Below is what is written in the Java source code:

```Java
...
     * Note that while some implementations of {@code InputStream} will return
     * the total number of bytes in the stream, many will not.  It is
     * never correct to use the return value of this method to allocate
     * a buffer intended to hold all data in this stream.
...
     * The {@code available} method for class {@code InputStream} always
     * returns {@code 0}.
...
    public int available() throws IOException {
        return 0;
    }
```

It might not be a bad idea to write your own version.

Hope it helps,
~p