@carl-mastrangelo 2016-08-17
Updated to reflect comments.  Also moved to using a loop for reading from the IS.