@pgrosu 2016-08-18
@carl-mastrangelo But does the Travis build pass now?  I waited an hour and they still haven't started, which might be due to the PR having been merged beforehand.